+++ src/collective/contact/core/browser/address.py
@@ -1,25 +1,33 @@
 from Acquisition import aq_base
 
 from five import grok
+from plone import api
 
 from collective.contact.core.behaviors import IContactDetails
 from collective.contact.core.browser import TEMPLATES_DIR
 from collective.contact.core.behaviors import ADDRESS_FIELDS
-
+from collective.contact.core.interfaces import IHeldPosition, IContactCoreParameters
 
 grok.templatedir(TEMPLATES_DIR)
 
 
 def get_address(obj):
     """Returns a dictionary which contains address fields"""
-    address = {}
-    if (aq_base(obj).use_parent_address is True and
-       hasattr(obj, 'aq_parent') and
-       IContactDetails.providedBy(obj.aq_parent)):
-        address = get_address(obj.aq_parent)
-        if address:
-            return address
+    if aq_base(obj).use_parent_address is True:
+        related = None
+        priv = api.portal.get_registry_record(name='person_contact_details_private', interface=IContactCoreParameters)
+        if IHeldPosition.providedBy(obj) and priv:
+            # For a held position, we use the related element: a position or an organization
+            related = (obj.get_position() or obj.get_organization())
+        elif hasattr(obj, 'aq_parent'):
+            related = obj.aq_parent
 
+        if related and IContactDetails.providedBy(related):
+            return get_address(related)
+        else:
+            return {}
+
+    address = {}
     address_fields = ADDRESS_FIELDS
     obj = aq_base(obj)
     for field in address_fields: