+++ src/collective/contact/core/browser/person.py
@@ -34,7 +34,7 @@ class HeldPositions(BrowserView):
         held_positions = []
         for obj in IPersonHeldPositions(person).get_sorted_positions():
             held_position = {}
-            held_position['label'] = obj.label or obj.Title()
+            held_position['title'] = obj.Title()
             if obj.start_date is not None:
                 start_date = date_to_DateTime(obj.start_date)
                 held_position['start_date'] = person.toLocalizedTime(start_date)
@@ -50,11 +50,12 @@ class HeldPositions(BrowserView):
             # held_position['phone'] = obj.phone
             # held_position['email'] = obj.email
             held_position['object'] = obj
-            # import pdb; pdb.set_trace()
-            if obj.get_organization():
-                held_position['organization'] = obj.get_organization().get_root_organization()
+            organization = obj.get_organization()
+            if organization:
+                held_position['organization'] = organization.get_root_organization()
             else:
-                continue
+                held_position['organization'] = None
+            held_position['icon'] = obj.getIconURL()
             held_position['can_edit'] = sm.checkPermission('Modify portal content', obj)
             held_position['can_delete'] = sm.checkPermission('Delete objects', obj)
             held_positions.append(held_position)