+++ src/collective/contact/core/subscribers.py
@@ -1,5 +1,6 @@
 from five import grok
 from z3c.form.interfaces import NO_VALUE
+
 from zope.container.contained import ContainerModifiedEvent
 from zope.lifecycleevent.interfaces import IObjectAddedEvent, \
     IObjectModifiedEvent
@@ -15,10 +16,10 @@ except ImportError:
         pass
 
 from collective.contact.core.behaviors import IContactDetails
-from collective.contact.core.content.held_position import IHeldPosition
 from collective.contact.core.content.position import IPosition
 from collective.contact.core.content.person import IPerson
 from collective.contact.core.content.organization import IOrganization
+from collective.contact.core.interfaces import IHeldPosition
 
 # update indexes of related content when a content is modified
 # you can monkey patch this value if you have an index that needs this