+++ src/collective/contact/core/tests/test_content_types.py
@@ -1,12 +1,14 @@
 # -*- coding: utf8 -*-
 
-import unittest2 as unittest
+import unittest
 
 import datetime
 
 from ecreall.helpers.testing.base import BaseTest
 
+from collective.contact.core.interfaces import IContactCoreParameters
 from collective.contact.core.testing import INTEGRATION
+from plone import api
 from plone.app.testing.interfaces import TEST_USER_ID, TEST_USER_NAME
 from plone.app.testing.helpers import setRoles
 
@@ -64,7 +66,7 @@ class TestPerson(TestContentTypes):
 
     def test_no_firstname(self):
         pepper = self.mydirectory['pepper']
-        self.assertEqual('Sergent Pepper', pepper.Title())
+        self.assertEqual('Mister Pepper', pepper.Title())
 
     def test_no_person_title(self):
         rambo = self.mydirectory['rambo']
@@ -167,6 +169,22 @@ class TestOrganization(TestContentTypes):
         self.mydirectory.manage_pasteObjects(cb)
         self.assertIn('copy_of_armeedeterre', self.mydirectory.keys())
 
+    def test_get_positions(self):
+        # add some positions to self.armeedeterre
+        self.armeedeterre.invokeFactory('position', 'colonel_adt', title="Colonel de l'armée de terre")
+        self.armeedeterre.invokeFactory('position', 'lieutenant_adt', title="Lieutenant de l'armée de terre")
+        self.armeedeterre.invokeFactory('position', 'sergent_adt', title="Sergent de l'armée de terre")
+        self.assertEquals(
+            [pos.id for pos in self.armeedeterre.get_positions()],
+            ['general_adt', 'colonel_adt', 'lieutenant_adt', 'sergent_adt'])
+        # get_positions sorts positions using getObjPositionInParent
+        # move 'general_adt' to last position
+        self.armeedeterre.moveObjectToPosition(
+            'general_adt', len(self.armeedeterre.objectIds()))
+        self.assertEquals(
+            [pos.id for pos in self.armeedeterre.get_positions()],
+            ['colonel_adt', 'lieutenant_adt', 'sergent_adt', 'general_adt'])
+
 
 class TestPosition(TestContentTypes):
 
@@ -182,7 +200,8 @@ class TestPosition(TestContentTypes):
         self.assertEqual(self.general_adt.get_full_title(),
                          u"Général de l'armée de terre (Armée de terre)")
         self.assertEqual(self.sergent_lh.get_full_title(),
-                         u"Sergent de la brigade LH, Brigade LH (Armée de terre)")
+                         u"Sergent de la brigade LH (Armée de terre / Corps A / Division Alpha / Régiment H / "
+                         u"Brigade LH)")
 
     def test_copy_paste(self):
         cb = self.armeedeterre.manage_copyObjects(['general_adt'])
@@ -200,26 +219,36 @@ class TestHeldPosition(TestContentTypes):
         sergent_pepper = self.sergent_pepper
         self.assertIn('adt', degaulle)
         self.assertEqual(adt.Title(),
-                         "Armée de terre")
+                         "(Armée de terre)")
         self.assertEqual(adt.title,
-                         "Armée de terre")
+                         u"(Armée de terre)")
         self.assertIn('gadt', degaulle)
         self.assertEqual(gadt.Title(),
-                         "Général de l'armée de terre (Armée de terre)")
+                         "Émissaire OTAN (Armée de terre)")
         self.assertIn('sergent_pepper', pepper)
         self.assertEqual(sergent_pepper.Title(),
-                         "Sergent de la brigade LH, Brigade LH (Armée de terre)")
+                         "Sergent de la brigade LH (Armée de terre / Corps A / Division Alpha / Régiment H / "
+                         "Brigade LH)")
         self.assertIsNone(sergent_pepper.end_date)
 
     def test_get_full_title(self):
         self.assertEqual(self.adt.get_full_title(),
                          u"Général Charles De Gaulle (Armée de terre)")
         self.assertEqual(self.gadt.get_full_title(),
-                         u"Général Charles De Gaulle (Armée de terre - Général de l'armée de terre)")
+                         u"Général Charles De Gaulle, Émissaire OTAN (Armée de terre)")
         self.assertEqual(self.sergent_pepper.get_full_title(),
-                         u"Sergent Pepper (Armée de terre - Sergent de la brigade LH)")
+                         u"Mister Pepper, Sergent de la brigade LH (Armée de terre / Corps A / Division Alpha / "
+                         u"Régiment H / Brigade LH)")
+
+    def test_get_person_title(self):
         self.assertEqual(self.gadt.get_person_title(),
                          u"Général Charles De Gaulle")
+        self.assertEqual(self.gadt.get_person_title(include_person_title=False),
+                         u"Charles De Gaulle")
+        api.portal.set_registry_record(
+            name='person_title_in_title', value=False, interface=IContactCoreParameters)
+        self.assertEqual(self.gadt.get_person_title(include_person_title=True),
+                         u"Charles De Gaulle")
 
     def test_get_person(self):
         pass